setter для валюты в документе расчет
