+++ package.json
@@ -26,12 +26,12 @@
     "history": "^4.7.2",
     "material-ui": "^1.0.0-beta.10",
     "material-ui-icons": "^1.0.0-beta.5",
-    "metadata-abstract-ui": "^2.0.2-beta.28",
-    "metadata-core": "^2.0.2-beta.28",
-    "metadata-dhtmlx": "^2.0.2-beta.28",
-    "metadata-pouchdb": "^2.0.2-beta.28",
-    "metadata-react": "^2.0.2-beta.28",
-    "metadata-redux": "^2.0.2-beta.28",
+    "metadata-abstract-ui": "^2.0.2-beta.29",
+    "metadata-core": "^2.0.2-beta.29",
+    "metadata-dhtmlx": "^2.0.2-beta.29",
+    "metadata-pouchdb": "^2.0.2-beta.29",
+    "metadata-react": "^2.0.2-beta.29",
+    "metadata-redux": "^2.0.2-beta.29",
     "moment": "^2.18",
     "paper-jsdom-canvas": "^0.11.4",
     "pouchdb": "^6.3.4",
